+++ core/utils.ts
@@ -274,6 +274,7 @@ export function getImportMap(userMap?: ImportMap, url?: URL): ImportMap {
       "lume": new URL("./mod.ts", baseUrl).href,
       "lume/": new URL("./", baseUrl).href,
       "https://deno.land/x/lume/": new URL("./", baseUrl).href,
+      "lume/plugins/date/locale": "https://deno.land/x/date_fns@v2.22.1/locale",
     },
   };