+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 - New plugin `netlify_cms`.
 - New `expires` middleware to add the `Expires` header to responses.
 - New `mergedKeys` data key to configure how some keys will be merged.
+- `lume/plugins/date/locale` bare import to the import_map.
 
 ### Changed
 - The `inline` plugin adds the `id` and `class` attributes